requirements.txt
# Tadawul Stock Collector - Dependencies
# Python 3.8+

# Data Collection
yfinance>=0.2.28              # Yahoo Finance API wrapper
requests>=2.31.0              # HTTP library for API calls

# Database
psycopg2-binary>=2.9.9        # PostgreSQL adapter
sqlalchemy>=2.0.23            # ORM and database toolkit

# Data Processing & Export
pandas>=2.1.3                 # Data manipulation and CSV export
numpy>=1.24.3                 # Numerical operations (pandas dependency)

# Configuration & Environment
python-dotenv>=1.0.0          # Environment variable management

# Progress & User Interface
tqdm>=4.66.1                  # Progress bars

# Date & Time
python-dateutil>=2.8.2        # Date parsing utilities

# Logging (built-in, but listing for completeness)
# logging - standard library